// src/app/models/vendor.model.ts
export interface Vendor {
  id?: string;
  vendorName: string;
  contactPerson: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  country: string;
  zipcode: string;
  phone1: string;
  phone2: string;
  fax: string;
  email1: string;
  email2: string;
  notes: string;
}